Styling

+++ app/components/buttons/ReservationButton.js
@@ -3,12 +3,13 @@ define(function() {
     var React = require('react');
     var TablesStore = require('./../stores/TablesStore');
     var ReservationsStore = require('./../stores/ReservationsStore');
-    var Modal = require('react-modal-component');
     var _ = require('lodash');
 
     var ReservationButton = React.createClass({
         propTypes: {
-            buttonType: React.PropTypes.string.isRequired
+            buttonType: React.PropTypes.string.isRequired,
+            toggleModal: React.PropTypes.func,
+            modalOpen: React.PropTypes.bool.isRequired
         },
 
         getInitialState: function() {
@@ -40,35 +41,51 @@ define(function() {
         },
 
         takeReservation: function() {
-            // TODO put in modal
-
+            if(!this.props.modalOpen) {
+                this.props.toggleModal();
+            }
         },
 
         removeReservation: function() {
-            var selectedReservation = _.find(this.state.reservations, ['selected', true]);
-            if(!selectedReservation) {
-                alert("Please select a reservation");
-            }
-            else {
-                var remove = confirm("Are you sure you want to delete this reservation?");
-                if(remove) {
-                    ReservationsStore.deleteReservation(selectedReservation.reservationId);
+            if(!this.props.modalOpen) {
+                var selectedReservation = _.find(this.state.reservations, ['selected', true]);
+                if (!selectedReservation) {
+                    alert("Please select a reservation");
+                }
+                else {
+                    var remove = confirm("Are you sure you want to delete this reservation?");
+                    if (remove) {
+                        ReservationsStore.deleteReservation(selectedReservation.reservationId);
+                    }
                 }
             }
         },
 
         seatReservation: function() {
-            var selectedTable = _.find(this.state.tables, ['selected', true]);
-            var selectedReservation = _.find(this.state.reservations, ['selected', true]);
-            if(!selectedTable || !selectedReservation) {
-                alert("Please select a table and reservation");
-            }
-            else {
-                if(selectedReservation.numberInParty > selectedTable.capacity) {
-                    var seat = confirm("The number in the selected party is \ngreater than the selected table capacity.\nContinue?");
+            //TODO change so multiple tables can be selected
+            if(!this.props.modalOpen) {
+                var selectedTable = _.find(this.state.tables, ['selected', true]);
+                var selectedReservation = _.find(this.state.reservations, ['selected', true]);
+                if (!selectedTable || !selectedReservation) {
+                    alert("Please select a table and reservation");
+                }
+                else {
+                    if (selectedReservation.numberInParty > selectedTable.capacity) {
+                        var seat = confirm("The number in the selected party is \ngreater than the selected table capacity.\nContinue?");
 
-                    if (seat) {
-                        // connect reservation to table
+                        if (seat) {
+                            // connect reservation to table
+                            selectedReservation.tableId = selectedTable.tableId;
+                            selectedReservation.selected = false;
+                            selectedReservation.display = false;
+                            selectedTable.party = selectedReservation;
+                            selectedTable.selected = false;
+                            selectedTable.occupied = true;
+                            TablesStore.setTableData(selectedTable.tableId, selectedTable);
+                            ReservationsStore.setReservationData(selectedReservation.reservationId, selectedReservation);
+                        }
+                    }
+                    else {
                         selectedReservation.tableId = selectedTable.tableId;
                         selectedReservation.selected = false;
                         selectedReservation.display = false;
@@ -78,16 +95,6 @@ define(function() {
                         TablesStore.setTableData(selectedTable.tableId, selectedTable);
                         ReservationsStore.setReservationData(selectedReservation.reservationId, selectedReservation);
                     }
-                }
-                else {
-                    selectedReservation.tableId = selectedTable.tableId;
-                    selectedReservation.selected = false;
-                    selectedReservation.display = false;
-                    selectedTable.party = selectedReservation;
-                    selectedTable.selected = false;
-                    selectedTable.occupied = true;
-                    TablesStore.setTableData(selectedTable.tableId, selectedTable);
-                    ReservationsStore.setReservationData(selectedReservation.reservationId, selectedReservation);
                 }
             }
         },
