+++ app/components/buttons/WaitlistButton.js
@@ -6,7 +6,9 @@ define(function() {
 
     var WaitlistButton = React.createClass({
         PropTypes: {
-            buttonType: React.PropTypes.string.isRequired
+            buttonType: React.PropTypes.string.isRequired,
+            toggleModal: React.PropTypes.func,
+            modalOpen: React.PropTypes.bool.isRequired
         },
 
         getInitialState: function() {
@@ -40,34 +42,51 @@ define(function() {
         },
 
         takeWaitlist: function() {
-            // TODO put in modal
+            if(!this.props.modalOpen) {
+                this.props.toggleModal();
+            }
         },
 
         removeWaitlist: function() {
-            var selectedWaitlist = _.find(this.state.waitlists, ['selected', true]);
-            if(!selectedWaitlist) {
-                alert("Please select a waitlist entry.");
-            }
-            else {
-                var remove = confirm("Are you sure you want to remove this waitlist entry?");
-                if(remove) {
-                    WaitlistsStore.deleteWaitlist(selectedWaitlist.waitlistId);
+            if(!this.props.modalOpen) {
+                console.log('here');
+                var selectedWaitlist = _.find(this.state.waitlists, ['selected', true]);
+                if (!selectedWaitlist) {
+                    alert("Please select a waitlist entry.");
+                }
+                else {
+                    var remove = confirm("Are you sure you want to remove this waitlist entry?");
+                    if (remove) {
+                        WaitlistsStore.deleteWaitlist(selectedWaitlist.waitlistId);
+                    }
                 }
             }
         },
 
         seatWaitlist: function() {
-            var selectedTable = _.find(this.state.tables, ['selected', true]);
-            var selectedWaitlist = _.find(this.state.waitlists, ['selected', true]);
-            if(!selectedTable || !selectedWaitlist) {
-                alert("Please select a table and waitlist entry");
-            }
-            else {
-                if(selectedWaitlist.numberInParty > selectedTable.capacity) {
-                    var seat = confirm("The number in the selected party is greater than the selected table capacity.\nContinue?");
+            if(!this.props.modalOpen) {
+                var selectedTable = _.find(this.state.tables, ['selected', true]);
+                var selectedWaitlist = _.find(this.state.waitlists, ['selected', true]);
+                if (!selectedTable || !selectedWaitlist) {
+                    alert("Please select a table and waitlist entry");
+                }
+                else {
+                    if (selectedWaitlist.numberInParty > selectedTable.capacity) {
+                        var seat = confirm("The number in the selected party is greater than the selected table capacity.\nContinue?");
 
-                    if (seat) {
-                        // connect reservation to table
+                        if (seat) {
+                            // connect reservation to table
+                            selectedWaitlist.tableId = selectedTable.tableId;
+                            selectedWaitlist.selected = false;
+                            selectedWaitlist.display = false;
+                            selectedTable.party = selectedWaitlist;
+                            selectedTable.selected = false;
+                            selectedTable.occupied = true;
+                            TablesStore.setTableData(selectedTable.tableId, selectedTable);
+                            WaitlistsStore.setWaitlistData(selectedWaitlist.reservationId, selectedWaitlist);
+                        }
+                    }
+                    else {
                         selectedWaitlist.tableId = selectedTable.tableId;
                         selectedWaitlist.selected = false;
                         selectedWaitlist.display = false;
@@ -77,16 +96,6 @@ define(function() {
                         TablesStore.setTableData(selectedTable.tableId, selectedTable);
                         WaitlistsStore.setWaitlistData(selectedWaitlist.reservationId, selectedWaitlist);
                     }
-                }
-                else {
-                    selectedWaitlist.tableId = selectedTable.tableId;
-                    selectedWaitlist.selected = false;
-                    selectedWaitlist.display = false;
-                    selectedTable.party = selectedWaitlist;
-                    selectedTable.selected = false;
-                    selectedTable.occupied = true;
-                    TablesStore.setTableData(selectedTable.tableId, selectedTable);
-                    WaitlistsStore.setWaitlistData(selectedWaitlist.reservationId, selectedWaitlist);
                 }
             }
         },
